Removed unneeded MeasurementCollection class.
