--- AntPlus/DeviceProfiles/BicyclePower/MeasurementCollection.cs
@@ -1,40 +1,0 @@
-﻿using System.Collections.ObjectModel;
-using System.Linq;
-
-namespace SmallEarthTech.AntPlus.DeviceProfiles.BicyclePower
-{
-    /// <summary>
-    /// A thread safe collection of measurements reported during calibration.
-    /// </summary>
-    public class MeasurementCollection : ObservableCollection<MeasurementOutputData>
-    {
-        /// <summary>
-        /// The collection lock.
-        /// </summary>
-        /// <remarks>
-        /// An application should use the collection lock to ensure thread safe access to the
-        /// collection. For example, the code behind for a WPF window should include -
-        /// <code>BindingOperations.EnableCollectionSynchronization(bicyclePowerViewModel.Bicycle.Calibration.Measurements, bicyclePowerViewModel.Bicycle.Calibration.Measurements.CollectionLock);</code>
-        /// This ensures changes to the collection are thread safe and marshalled on the UI thread.
-        /// </remarks>
-        public object CollectionLock = new object();
-
-        /// <summary>
-        /// Parses the specified data page.
-        /// </summary>
-        /// <param name="dataPage">The data page.</param>
-        internal void Parse(byte[] dataPage)
-        {
-            MeasurementOutputData measurement = this.FirstOrDefault(m => m.MeasurementType == (MeasurementOutputData.DataType)dataPage[2]);
-            if (measurement == null)
-            {
-                measurement = new MeasurementOutputData(dataPage);
-                lock (CollectionLock)
-                {
-                    Add(measurement);
-                }
-            }
-            measurement.Parse(dataPage);
-        }
-    }
-}
